Move test to there own file
